// packages/sensinact-mapping-wizard/src/generated/StoragePreset.ts
/*
 * This file was generated by emfts-codegen.
 * DO NOT EDIT - changes will be overwritten.
 *
 * @generated
 */

/**
 * StoragePreset
 * @generated
 */
export const StoragePreset = {
  EVERY_CHANGE: 'EVERY_CHANGE',
  CHANGED_5_PERCENT: 'CHANGED_5_PERCENT',
  MAX_ONCE_10MIN: 'MAX_ONCE_10MIN'
} as const;

export type StoragePreset = typeof StoragePreset[keyof typeof StoragePreset];

/**
 * Get enum value by name
 */
export function getStoragePreset(name: string): StoragePreset | undefined {
  return (StoragePreset as Record<string, string>)[name] as StoragePreset | undefined;
}

/**
 * Get enum value by literal
 */
export function getStoragePresetByLiteral(literal: string): StoragePreset | undefined {
  if (literal === 'EVERY_CHANGE') return StoragePreset.EVERY_CHANGE;
  if (literal === 'CHANGED_5_PERCENT') return StoragePreset.CHANGED_5_PERCENT;
  if (literal === 'MAX_ONCE_10MIN') return StoragePreset.MAX_ONCE_10MIN;
  return undefined;
}